Refactoring the Dashboard page code
