+++ client/src/pages/auth/dashboard/components/QuickActions.jsx
@@ -1,0 +1,33 @@
+import "./QuickActions.css";
+
+export default function QuickActions({
+  openNewTaskPopup,
+  openNewProjectPopup,
+}) {
+  return (
+    <div className="quick-actions">
+      <h4 className="title poppins-bold">Quick actions</h4>
+      <ul className="actions poppins-regular">
+        <li className="action">
+          <button
+            className="poppins-regular"
+            onClick={() => openNewTaskPopup()}
+          >
+            Create a new task
+          </button>
+        </li>
+        <li className="action">
+          <button
+            className="poppins-regular"
+            onClick={() => openNewProjectPopup()}
+          >
+            Create a new project
+          </button>
+        </li>
+        <li className="action feature-disabled">
+          <button className="poppins-regular">Complete today's tasks</button>
+        </li>
+      </ul>
+    </div>
+  );
+}
